Expanded upon optional grid system and alignments, restyling
This adds the options for the screen and extra-large grid formats
optionally. They can be turned on if desired for further flexibility,
but are not considered to be part of the "core functionality".

The Goal with this is to allow customization for larger screens, and
more precision on smaller screens. However, due to the amount of
queries required to touch each screen size, they are disabled by
default. This not only creates shorter css output, but encourages
less breakpoints while leaving it available for needed cases.

This update also adds the uncentered option. The reason for this is
in the case where you are centering a div on a smaller screen, the
larger screen will be effected. This allows a reset of that from that
media query and up (or at least it should).